End of section 5 - add Identity Service

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -50,13 +51,13 @@ public class AuctionsController : ControllerBase
     return _mapper.Map<AuctionDto>(auction);
   }
 
+  [Authorize]
   [HttpPost]
   public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
   {
     var auction = _mapper.Map<Auction>(auctionDto);
-    // TODO: add current user as seller
 
-    auction.Seller = "test";
+    auction.Seller = User.Identity.Name;
 
     _context.Auctions.Add(auction);
     var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -69,6 +70,7 @@ public class AuctionsController : ControllerBase
     return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, _mapper.Map<AuctionDto>(auction));
   }
 
+  [Authorize]
   [HttpPut("{id}")]
   public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
   {
@@ -77,7 +79,7 @@ public class AuctionsController : ControllerBase
 
     if (auction == null) return NotFound();
 
-    // TODO: check seller
+    if (auction.Seller != User.Identity.Name) return Forbid();
 
     auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
     auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -94,6 +96,7 @@ public class AuctionsController : ControllerBase
     return BadRequest("Problem when saving to DB");
   }
 
+  [Authorize]
   [HttpDelete("{id}")]
   public async Task<ActionResult> DeleteAuction(Guid id)
   {
@@ -101,11 +104,11 @@ public class AuctionsController : ControllerBase
 
     if (auction == null) return NotFound();
 
-    // TODO: check seller == username
+    if (auction.Seller != User.Identity.Name) return Forbid();
 
     _context.Auctions.Remove(auction);
 
-    await _publishEndpoint.Publish<AuctionDeleted>(new {Id = auction.Id.ToString()});
+    await _publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
 
     var result = await _context.SaveChangesAsync() > 0;
